requirements.txt
sounddevice==0.5.1

# Paquets utilisé en environement virtualisé de python
rpi-lgpio==0.6 
pygame==2.6.1
numpy==2.1.3
pydub==0.25.1
#A des fin de test, et seulement de test (pour appuyer sur la touche quiter.)


# Les paquets ci-dessous sont déjà installé par défault, je les places en commentaire.
#TODO: a tester avec une raspberry reformaté
#lgpio==0.2.2.0
#rpi-lgpio==0.6
#RPi.GPIO==0.7.1
#numpy==2.1.3